Move PayButton to :core
It is not a pure UI component and will need string resources, so :core is a better place.

COSDK-1312
