+++ core/src/main/java/com/adyen/checkout/core/components/internal/ui/PayButton.kt
@@ -1,17 +1,18 @@
 /*
- * Copyright (c) 2025 Adyen N.V.
+ * Copyright (c) 2026 Adyen N.V.
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by oscars on 5/12/2025.
+ * Created by oscars on 13/7/2026.
  */
 
-package com.adyen.checkout.ui.internal.element.button
+package com.adyen.checkout.core.components.internal.ui
 
 import androidx.annotation.RestrictTo
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import com.adyen.checkout.ui.internal.element.button.PrimaryButton
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @Composable
